Refactor services into dedicated package
